oops; minor bug in cleaning up old validations
The effect is that people who revalidate before they need to, will have
several wasteful records in ~/.ssh/authorized_keys.

On the plus side, this helped me catch someone who was programmatically
revalidating himself every 15 minutes, due to the number of leftover
directories with his name in them.  This means he had the second factor on his
laptop.  He's going to get it in the neck tomorrow morning...

(I store the secret on my laptop too, for emergencies, but it's gpg protected.
And gpg-agent's ttl's are set pretty low -- far lower than he seems to have,
if he is using it at all!)